Work around lack of exported S3 summary method for mtc.result
The function extractResults in this package requires access to the S3 method for the summary generic objects of class mtc.result from gemtc. summary is not exported from the gemtc namespace so added a function that uses getS3method to access the required method. Probably not best practice but the best I can do.